requirement.txt
flask 
google-generativeai 
requests